Added missing packages for the linux installer...
Also reenabled monodevelop and got it working

Added some error checking

Create a list of rpms to put in the readme if the readme for the release
doesn't exists. (traceability)  Because really, the installer will get
the latest of what's out there, not necessarily what's been release.

Added some subversion properties to ignore build-time files



svn path=/trunk/release/; revision=45022